Update EthereumLikePaymentData.cs
switched out ulong for BigInteger to handle larger value mapping.

+++ BTCPayServer/Services/Altcoins/Ethereum/Payments/EthereumLikePaymentData.cs
@@ -10,7 +10,7 @@ namespace BTCPayServer.Services.Altcoins.Ethereum.Payments
 {
     public class EthereumLikePaymentData : CryptoPaymentData
     {
-        public ulong Amount { get; set; }
+        public BigInteger Amount { get; set; }
         public string CryptoCode { get; set; }
         public string Address { get; set; }
         public long AccountIndex { get; set; }
